refactor(archive): rework single-export files
